requirements.txt
flask
flask-socketio
opencv-python
mediapipe
protobuf<4.0.0
numpy
Pillow
eventlet
